requirements.txt
flask
flask-cors
flask-swagger
swagger_ui_bundle
gunicorn
pyyaml
